+++ plone/app/users/browser/membersearch.py
@@ -7,8 +7,6 @@ from plone.supermodel import model
 from z3c.form import button
 from z3c.form import form
 from zope import schema
-# from plone.autoform import directives
-# from z3c.form.browser.checkbox import CheckBoxWidget
 
 
 class IMemberSearchSchema(model.Schema):